Update the start script to use the version functions
